Construct ping URL preserving scheme
Downtimer is used to create a ping URL by extracting `host` and `port`
components out of URL, and combining them with `http://` prefix. The
approach is error prone as OpenStack deployments are not necessarily
allows http:// access to their endpoints. For instance, OpenStack
Ansible is https:// only by default.

So this commit preserves the original scheme of the URL.